+++ launcher/src-tauri/src/core/loaders/pvp.rs
@@ -21,13 +21,13 @@ const GITHUB_REPO: &str = "SantiJ10/Paraguacraft";
 const MANIFEST_URL: &str =
     "https://raw.githubusercontent.com/SantiJ10/Paraguacraft/main/clientes/paraguacraft-pvp/manifest.json";
 
-const FALLBACK_CLIENT_VERSION: &str = "2.1.15";
-const FALLBACK_RELEASE_TAG: &str = "pvp-client-2.1.15";
+const FALLBACK_CLIENT_VERSION: &str = "2.1.16";
+const FALLBACK_RELEASE_TAG: &str = "pvp-client-2.1.16";
 
 const FALLBACK_MODS: &[(&str, &str)] = &[
     (
-        "ParaguacraftPvP-2.1.15.jar",
-        "9fe263f0902750f564e5dfd4f62b195217fe3443",
+        "ParaguacraftPvP-2.1.16.jar",
+        "3e6e47607ef87de0271f6e02604c1312b805e504",
     ),
     (
         "OptiFine_1.8.9_HD_U_M5.jar",